Move Slider from react-slider to react-internal
