+++ packages/react-examples/src/react/Slider/Slider.Vertical.Example.tsx
