+++ packages/react-examples/src/react/Slider/docs/SliderBestPractices.md
