+++ packages/react-internal/src/components/Slider/Slider.base.tsx
