+++ packages/react-internal/src/components/Slider/Slider.styles.ts
