+++ packages/react-internal/src/components/Slider/Slider.types.ts
