+++ packages/react-internal/src/components/Slider/index.ts
@@ -1,0 +1,3 @@
+export * from './Slider';
+export * from './Slider.base';
+export * from './Slider.types';
