+++ packages/react-slider/src/components/Slider/Slider.scss
