--- packages/react-slider/src/next/index.ts
@@ -1,2 +1,0 @@
-export * from '../index';
-export { Slider } from './Slider';
